Move PvaEpicsGUI to pva/gui.py
